Added basic bash configuration files
- added bashrc
- added bash_profile
- setup.sh, added support for bashrc and bash_profile

+++ setup.sh
@@ -1,15 +1,20 @@
 #!/bin/bash
 
-#--------- VIM ---------#
+#--------- Bash ---------#
 
-# Check if .vimrc exitst and create a backup
-if [ -f ~/.vimrc ]; then
-	cp ~/.vimrc ~/.vimrc.bk
-fi
-if [ -d ~/.vim ]; then
-	rm -rf ~/.vim-bk
-	mv ~/.vim ~/.vim-bk
-fi
+# Backup bash files
+[[ -f ~/.bash_profile ]] && cp ~/.bash_profile ~/.bash_profile.bk
+[[ -f ~/.bashrc ]] && cp ~/.bashrc ~/.bashrc.bk
+
+cp bash_profile ~/.bash_profile
+cp bashrc ~/.bashrc
+
+
+#--------- VIM ----------#
+
+# Check if .vimrc and .vim exitst and create a backup
+[[ -f ~/.vimrc ]] && cp ~/.vimrc ~/.vimrc.bk
+[[ -d ~/.vim ]] && rm -rf ~/.vim-bk && mv ~/.vim ~/.vim-bk
 
 # Copy vimrc and folders (if any)
 cp vimrc ~/.vimrc
@@ -19,9 +24,7 @@ cp -R vim ~/.vim
 #--------- Screen ---------#
 
 # Check if .vimrc exitst and create a backup
-if [ -f ~/.screenrc ]; then
-	cp ~/.screenrc ~/.screenrc.bk
-fi
+[[ -f ~/.screenrc ]] && cp ~/.screenrc ~/.screenrc.bk
 
 # Copy screenrc
 cp screenrc ~/.screenrc
